Create more tasks and attempt to separate dependencies
Must be tested further
