+++ server/src/utils/image.py
@@ -4,7 +4,7 @@ import numpy as np
 import glob
 import logging as log
 import hdbscan
-from src.utils.export import get_file_basename
+from src.utils.file import get_file_basename
 from src.utils.file import save_file_layouts
 
 ##################################################
